♻️ Refactor the ubication of interfaces
